Code cleanup - restructuring component files for easier readability
