+++ client/components/helpers/editor_preview.jsx
