+++ client/components/routes.jsx
@@ -9,22 +9,22 @@ import { BrowserRouter as Router, Switch } from "react-router-dom";
 
 // --------------------------------------------------------------------------------------------------------------------------------------------
 // Import all components for navigation
-import App from "./components/app.jsx";
-import Home from "./components/home.jsx";
-import Login from "./components/login.jsx";
-import Article from "./components/article.jsx";
-import NewArticle from "./components/new.jsx";
-import EditArticle from "./components/edit.jsx";
-import ArticleHistory from "./components/history.jsx";
-import Search from "./components/search.jsx";
-import Admin from "./components/admin.jsx";
-import Setup from "./components/setup.jsx";
-import EditTopic from "./components/edit_topics.jsx";
-import EditUser from "./components/edit_users.jsx";
-import UserSignup from "./components/user_signup.jsx";
-import Landing from "./components/landing.jsx";
-import About from "./components/about.jsx";
-import Results from "./components/results.jsx";
+import App from "./main/app.jsx";
+import Home from "./main/home.jsx";
+import Login from "./users/login.jsx";
+import Article from "./wiki/article.jsx";
+import NewArticle from "./wiki/new.jsx";
+import EditArticle from "./wiki/edit.jsx";
+import ArticleHistory from "./wiki/history.jsx";
+import Search from "./search/search.jsx";
+import Admin from "./users/admin.jsx";
+import Setup from "./users/setup.jsx";
+import EditTopic from "./users/edit_topics.jsx";
+import EditUser from "./users/edit_users.jsx";
+import UserSignup from "./users/user_signup.jsx";
+import Landing from "./main/landing.jsx";
+import About from "./main/about.jsx";
+import Results from "./search/results.jsx";
 
 // --------------------------------------------------------------------------------------------------------------------------------------------
 // Build a route path and limit access to certain pages
