+++ client/components/search/filters.jsx
@@ -2,7 +2,7 @@
   Logic to handle automatic artwork filtering inside the search results page
 */
 import React from "react";
-import Loader from "./loader.jsx";
+import Loader from "../helpers/loader.jsx";
 //import Alert from "react-s-alert";
 
 class Filters extends React.Component {
