+++ client/components/users/edit_topics.jsx
@@ -4,7 +4,7 @@
 import React from "react";
 import { hashHistory } from "react-router";
 //import Alert from "react-s-alert";
-import Loader from "./loader.jsx";
+import Loader from "../helpers/loader.jsx";
 
 class EditTopic extends React.Component {
   constructor(props) {
