+++ client/components/users/edit_users.jsx
@@ -4,7 +4,7 @@
 import React from "react";
 import { hashHistory } from "react-router";
 //import Alert from "react-s-alert";
-import Loader from "./loader.jsx";
+import Loader from "../helpers/loader.jsx";
 
 class EditUser extends React.Component {
   constructor(props) {
