+++ client/components/wiki/article.jsx
@@ -4,7 +4,7 @@
 */
 import React from "react";
 import { Link, hashHistory } from "react-router";
-import Loader from "./loader.jsx";
+import Loader from "../helpers/loader.jsx";
 
 //import Alert from "react-s-alert";
