+++ client/components/wiki/browse_articles.jsx
@@ -3,7 +3,7 @@
   to engage platform users.
 */
 import React from "react";
-import Loader from "./loader.jsx";
+import Loader from "../helpers/loader.jsx";
 import { Link, hashHistory } from "react-router";
 //import Alert from "react-s-alert";
