+++ client/components/wiki/simple_article.jsx
@@ -3,7 +3,7 @@
 */
 import React from "react";
 import { Link } from "react-router";
-import Loader from "./loader.jsx";
+import Loader from "../helpers/loader.jsx";
 //import Alert from "react-s-alert";
 
 class SimpleArticle extends React.Component {
